separated polynomial, removed excessive doctests
